Added basic menu + player now has to type in server IP to join.
multiple clients on different pcs on same network can player together

+++ Client/include/Tile.h
@@ -2,5 +2,5 @@
 
 struct Tile
 {
-    int type = 0; //tile ID used for world rendering
+    uint16_t type = 0; //tile ID used for world rendering
 };
